refactor: useTypeInput -> useInput 으로 변경
